Move certificate file splitting to common
